Added some basic gameplay mechanics and tests (#33)
* Added some basic gameplay mechanics and tests

Created some basic mechanics for the game so the UI can start to take shape. Also added a few tests to ensure the simplest things can work.

Gameplay is designed around a board instance and a few other constructs to enable easy code structure. No computer player or anything yet, just a wrapper for the board array and some helper methods.

* Add a whole bunch more gameplay

* Fix project organization a bit

There are a few more projects, so the naming can't be app and then full project names.

* Build Debug for now

* Update README.md